types

+++ hypershot_uncertainty.py
@@ -97,7 +97,7 @@ def experiment(N):
     # working on it in a hurry
     ims = get_image_size(params)
     bs =  model.n_way*(model.n_support + model.n_query)*ims*ims
-    bn = torch.numel(X)/bs
+    bn = int(torch.numel(X)/bs)
     print(bn)
     B = torch.reshape(X, (bn, model.n_way*(model.n_support + model.n_query), ims, ims))
